Extract CircularBuffer<T> to separate file
Moves CircularBuffer<T> generic class from DebugOverlayController.cs to its own file following Single Responsibility Principle.

Changes:
- Create CircularBuffer.cs in Runtime/Core/ (71 lines)
- Remove CircularBuffer class from DebugOverlayController.cs (345 → 279 lines)
- Preserve namespace and all functionality
- No changes to tests (namespace import already works)

Benefits:
- Improves discoverability and reusability
- Enables easier unit testing in isolation
- Follows one-class-per-file convention
- Reduces DebugOverlayController.cs complexity
